Fixed LeastSquaedRegression errors, still need to fix graphing

+++ WheelSpeedCalibration/src/wheelspeedcalibration/NormalizePowerValues.java
@@ -10,20 +10,21 @@ import java.util.ArrayList;
  *
  * @author Adam
  */
-public class NormalizePowerValues 
+public class NormalizePowerValues
 {
     // double indexVal = p.index
     // Normalized Value = -1 + (indexVal /(arraylist.size() / 2))
+
     public static DoubleArrayList normalizeValues(ArrayList<GraphPoint> arrayList)
     {
         DoubleArrayList dblArrayList = new DoubleArrayList();
         ArrayList<GraphPoint> posArrayList = new ArrayList<>();
         ArrayList<GraphPoint> negArrayList = new ArrayList<>();
         double indexVal;
-        for (GraphPoint p: arrayList)
-        {            
-            indexVal = p.index;
-            p.power = -1 + (indexVal / (arrayList.size() / 2));         
+        for (GraphPoint p : arrayList)
+        {
+            indexVal = (double) (p.index);
+            p.power = -1 + (indexVal / ((double)(arrayList.size()) / 2.0));
             if (p.power <= 0)
             {
                 posArrayList.add(p);
